Move download-dataset script out of training
It is not used in training anymore, it is just a utility script for
admins to get TSVs files to pass to genie manually.

And as such, expose it as an almond-cloud entrypoint.
